% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{cv.grid_search}
\alias{cv.grid_search}
\title{Grid search to find parameters giving best cross-validated score}
\usage{
cv.grid_search(setup, f, ..., maximize = TRUE, nested = FALSE)
}
\arguments{
\item{setup}{cv settings created by cv.setup}

\item{f}{function that trains a model and makes prediction. Its 4 first inputs have to be xtrain, ytrain, xtest, ytest. Should return predictions of xtest or final score, depending on cv setup, see details.}

\item{...}{grid in the form \code{logreg = -2:3} where logreg is input parameter to f}

\item{maximize}{whether to maximize or minimize the score (first of the scores if cv.setup has several scores)}

\item{nested}{whether to perform nested CV}
}
\description{
Grid search to find parameters giving best cross-validated score
}
\examples{
x <- matrix(runif(50*5), 50, 40)
y <- x[,1] + 0.5*x[,2] + 0.1*runif(50)

## linear regression
regr <- function(x, y, xtest, ytest, logreg=0) {
  C =  diag(x = 10^logreg, ncol(x))
  beta = solve(t(x) \%*\% x + C, t(x) \%*\% y)
  xtest \%*\% beta
}
cv <- cv.setup(x, y, score=mean_se, num_folds = 5, num_iter = 2)
res <- cv.grid_search(cv, regr, logreg = -2:3, maximize=FALSE )
}
\seealso{
\code{\link{cv.run}} for how to define learn-predict function f.
}

