% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grid_search}
\alias{grid_search}
\title{Finds optimum for function using grid search.}
\usage{
grid_search(f, ..., maximize = TRUE)
}
\arguments{
\item{f}{function to be optimized}

\item{...}{grid in the form x = c(-5, -1, 1, 5) where x is input parameter to f}

\item{maximize}{whether to maximize or minimize}
}
\value{
solution and details
}
\description{
Finds optimum for function using grid search.
}
\examples{
f <- function(x,y) -x*x - 0.5*y*y
opt <- grid_search(f, x=seq(-5, 5, 2), y=seq(-5, 5, 2))
## solution found
opt$solution
## value of f at the solution
opt$optimum
}
\seealso{
\code{\link{random_search}}, \code{\link{nelder_mead}} and \code{\link{particle_swarm}} for other high level methods for optimization.
}

