% Generated by roxygen2 (4.0.2): do not edit by hand
\name{manual}
\alias{manual}
\title{Displays manuals for optimization methods in Optunity package.}
\usage{
manual(solver_name = "")
}
\arguments{
\item{solver_name}{solver name, or an empty string for a of all available solvers}
}
\value{
solver name(s) or list of available solvers if solver_name=''
}
\description{
Displays manuals for optimization methods in Optunity package.
}
\examples{
manual('')
manual('particle_swarm')
}
\seealso{
\code{\link{optimize2}} low level function for running different solvers
}

