% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nelder_mead}
\alias{nelder_mead}
\title{Finds optimum for function using Nelder-Mead.}
\usage{
nelder_mead(f, ..., num_evals = 50, maximize = TRUE)
}
\arguments{
\item{f}{function to be optimized}

\item{...}{starting point in the form x = 5, where x is input parameter to f}

\item{num_evals}{maximum number of evaluations of f}

\item{maximize}{whether to maximize or minimize}
}
\value{
solution and details
}
\description{
Finds optimum for function using Nelder-Mead.
}
\details{
Nelder-Mead is efficient if f is convex, otherwise it can easily get stuck into local minima.
}
\examples{
f <- function(x,y) -x*x - 0.5*y*y
opt <- nelder_mead(f, x=5, y=5)
## solution found
opt$solution
## value of f at the solution
opt$optimum
}
\seealso{
\code{\link{grid_search}}, \code{\link{random_search}}, and \code{\link{particle_swarm}} for other high level methods for optimization.
}

