% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optimize2}
\alias{optimize2}
\title{Low level function to call any optimizer with custom configuration.}
\usage{
optimize2(f, solver_name, solver_config = list(), constraints = NULL,
  maximize = TRUE, max_evals = 0, call_log = NULL, default = NULL)
}
\arguments{
\item{f}{function to be optimized}

\item{solver_name}{name of the solver (see manual() for available solvers)}

\item{solver_config}{list of configuration options for the solver}

\item{constraints}{list of constraints, if needed}

\item{maximize}{whether to maximize (TRUE) or minimize (FALSE)}

\item{max_evals}{number of evaluations to perform}

\item{call_log}{currently not used.}

\item{default}{default value given to solvers when constraints are violated}
}
\value{
result of optimization
}
\description{
Low level function to call any optimizer with custom configuration.
}
\examples{
## function to be minimized
f <- function(x,y,z) { (x-1)^2 + (y-1)^2 + (z-2)^2 }
## bounds for parameters
args = list(x = c(-10, 10), y = c(-10, 10), z = c(-10, 10))
## particle swarm setup
args$num_particles   = 8
args$num_generations = 10

optimize2(f, solver_name="particle swarm", maximize=FALSE, solver_config =args )
}
\seealso{
\code{\link{grid_search}}, \code{\link{random_search}}, \code{\link{nelder_mead}} and \code{\link{particle_swarm}} for high level methods for optimization.
}

