% Generated by roxygen2 (4.0.2): do not edit by hand
\name{particle_swarm}
\alias{particle_swarm}
\title{Finds optimum for function using Particle Swarm.}
\usage{
particle_swarm(f, ..., num_particles = 5, num_generations = 10,
  maximize = TRUE)
}
\arguments{
\item{f}{function to be optimized}

\item{...}{box constraints of form x = c(-5, 5) where x is input parameter to f}

\item{num_particles}{number of particles}

\item{num_generations}{number of generations}

\item{maximize}{whether to maximize or minimize}
}
\value{
solution and details
}
\description{
Finds optimum for function using Particle Swarm.
}
\examples{
f <- function(x,y) -x*x - 0.5*y*y
opt <- particle_swarm(f, x=c(-5, 5), y=c(-5, 5) )
## solution found
opt$solution
## value of f at the solution
opt$optimum
}
\seealso{
\code{\link{grid_search}}, \code{\link{random_search}}, and \code{\link{particle_swarm}} for other high level methods for optimization.
}

