% Generated by roxygen2 (4.0.2): do not edit by hand
\name{random_search}
\alias{random_search}
\title{Finds optimum for function using random search.}
\usage{
random_search(f, ..., maximize = TRUE, num_evals = 50)
}
\arguments{
\item{f}{function to be optimized}

\item{...}{box constraints of form x = c(-5, 5) where x is input parameter to f}

\item{maximize}{whether to maximize or minimize}

\item{num_evals}{maximum number of evaluations of f}
}
\value{
solution and details
}
\description{
Finds optimum for function using random search.
}
\examples{
f <- function(x,y) -x*x - 0.5*y*y
opt <- random_search(f, x=c(-5,5), y=c(-5,5), num_evals=40 )
## solution found
opt$solution
## value of f at the solution
opt$optimum
}
\seealso{
\code{\link{grid_search}}, \code{\link{nelder_mead}} and \code{\link{particle_swarm}} for other high level methods for optimization.
}

