% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Scoring}
\alias{Scoring}
\alias{accuracy}
\alias{accuracy_from_p}
\alias{auc_pr}
\alias{auc_roc}
\alias{early_bedroc}
\alias{early_rie}
\alias{mean_ae}
\alias{mean_se}
\title{Scoring functions for measuring model performance}
\usage{
mean_se(ytrue, yhat)

mean_ae(ytrue, yhat)

accuracy_from_p(ytrue, yprob)

accuracy(ytrue, yhat)

auc_roc(ytrue, yscore, decreasing = TRUE, top = 1)

auc_pr(ytrue, yscore, decreasing = TRUE)

early_rie(ytrue, yscore, decreasing = TRUE, alpha = 20)

early_bedroc(ytrue, yscore, decreasing = TRUE, alpha = 20)
}
\arguments{
\item{ytrue}{true (test) values for y}

\item{yhat}{estimated values for y}

\item{yprob}{estimated probabilities of positive class, used for binary classification}

\item{yscore}{ranking scores for computing AUC for ROC and Precision-Recall curve}

\item{decreasing}{whether high score corresponds to TRUE (positive) class or FALSE (negative) class}

\item{top}{value between 0.0 and 1.0, the proportion of top ratings used, if 1.0 all are used for computing AUC, if 0.1 then AUC only depends on top 10\%}

\item{alpha}{how many of the top ranked samples are important, used in early discovery metrics to define what is considered 'early', default 20.}
}
\value{
mean of squared errors between ytrue and yhat

mean of absolute errors between ytrue and yhat

proportion of correct predictions with yprob > 0.5 is used as cut-off

proportion of correct predictions, i.e. \code{mean(ytrue==yhat)}

Area under ROC curve

Area under precision-recall curve

Early discovery metric RIE (robust initial enhancement)

Early discovery metric BEDROC (Boltzmann-enhanced discrimination of ROC)
}
\description{
Scoring functions for measuring model performance
}

