% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_solver}
\alias{make_solver}
\title{Tests if solver can be made with supplied arguments. Low level API.}
\usage{
make_solver(solver_name, ...)
}
\arguments{
\item{solver_name}{solver name}

\item{...}{parameters passed to making the solver}
}
\value{
TRUE if succeeds, otherwise throws error
}
\description{
Tests if solver can be made with supplied arguments. Low level API.
}
\examples{
make_solver('random search', num_evals = 30)
}
\seealso{
\code{\link{manual}} for list of solvers and their arguments
}

